backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
python-multipart==0.0.6
torch
numpy
Pillow
python-dotenv
aiofiles

